requirements.txt
ipaddress==1.0.16
ipfs-api==0.2.2
python-geoip==1.2
python-geoip-geolite2==2015.303
requests==2.9.1
six==1.10.0
wheel==0.24.0
